Insert "time in test" column to log message below [JFG-522] (round)

+++ chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/scenario/WorkloadExecutionStatusBuilder.java
@@ -154,7 +154,7 @@ public class WorkloadExecutionStatusBuilder {
                 report += String.format(format,
                         node.getIdentifier(), this.threads.get(node),
                         this.startedSamples.get(node),
-                        this.finishedSamples.get(node), this.delays.get(node), (double)this.durationTime.get(node)/1000.0);
+                        this.finishedSamples.get(node), this.delays.get(node), this.durationTime.get(node)/1000);
             }
             return report + line;
         }
